fix iOS Safari centering issue
The problem occurs when location is blocked and the user searches for a
location. On scroll, the map refreshes and the map centers to the
original declared center. The map’s last searched center should persist
when the user scrolls.

So, doubling up on a couple of things seems to resolve the issue and
the last search center point persists even when the user scrolls away
from the map on iOS Safari. Latitude and Longitude are changed to
global variables and a map search gives the variables new values. This,
combined with removing the calculateCenter function resolves the issue.